% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_forecast.R
\name{download_forecast}
\alias{download_forecast}
\title{Download Gage}
\usage{
download_forecast(gage_id, path)
}
\arguments{
\item{gage_id}{character; The 5-digit river gage identifer. The list
can be found at https://water.weather.gov/ahps/.}

\item{path}{character; The path to the folder where the `.xml` file
will be written to.}
}
\value{
Path the downloaded gage forcast .xml file.
}
\description{
Download the `.xml` file for the specified gage from the NWS
AHPS web service.
}
