% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gage_forecast.R
\name{get_gage_forecast}
\alias{get_gage_forecast}
\title{Get Gage Forecast Data}
\usage{
get_gage_forecast(gage_id, gage_file)
}
\arguments{
\item{gage_id}{character; The 5-digit river gage identifer. The list
can be found at https://water.weather.gov/ahps/.}

\item{gage_file}{character; The path to a gage forecast xml file.}
}
\value{
data frame; A data frame containing gage site information for the
  input gage forecast xml file.
}
\description{
Extract the gage forecast information from the input gage xml
  file and write it into a data frame.
}
