% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gage_observed.R
\name{get_gage_observed}
\alias{get_gage_observed}
\title{Get Gage Observed Date}
\usage{
get_gage_observed(gage_id, gage_file)
}
\arguments{
\item{gage_id}{character; The 5-digit river gage identifer. The list
can be found at https://water.weather.gov/ahps/.}

\item{gage_file}{character; The path to a gage forecast xml file.}
}
\value{
data frame; A data frame containing gage site information for the
  input gage forecast xml file.
}
\description{
Extract the gage observed information from the input gage xml
  file and write it into a data frame.
}
\examples{
gage_id <- "canm7"
gage_file <- file.path(system.file("extdata", "canm7.xml",
                                   package = "rahps"))

gage_observed_df <- get_gage_observed(gage_id, gage_file)

}
