% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_forecast.R
\name{download_forecast}
\alias{download_forecast}
\title{Download Gage}
\usage{
download_forecast(gage_id, path)
}
\arguments{
\item{gage_id}{character; The 5-digit river gage identifier. The list
can be found at https://water.weather.gov/ahps/.}

\item{path}{character; The path to the folder where the `.xml` file
will be written to.}
}
\value{
character; The path to the downloaded gage forecast xml file.
}
\description{
Download the `.xml` file for the specified gage from the NWS
AHPS web service.
}
\details{
This function uses the National Weather Service (NWS) Advanced
Hydrologic Prediction Service (AHPS) `hydrograph_to_xml` web service to
download gage forcasts, using a URL such as:
`https://water.weather.gov/ahps2/hydrograph_to_xml.php?gage=blvi4&output=xml`
}
\examples{
\donttest{
gage_id <- "canm7"
path <- Sys.getenv("HOME")

gage_file <- rahps::download_forecast(gage_id, path)

# View the xml file
htmltidy::xml_view(xml2::read_xml(gage_file))
}

}
